Added stack build files to the project
Transitioned the project from building it directly using cabal to
building through the stack interface. For now, we opted for the current
LTS resolver, which is 18.13 (using ghc-8.10.7).

The most notable changes are:

- Renamed the directory where the library files are housed from "lib" to
"src". This seems to be the default (and kind of a loose convention)
using stack, so let's just stick to it.

- Added a stack.yaml file. This file is not 100% ready (there are some
project metadata fields that are still on the defaults, we will work
on it later).

- Added a package.yaml file. This file is now, according to stack
documentation, where we should deal with our packages, instead of
editing the Pruvarilo.cabal file directly. Currently, all package
dependencies used were found in stackage with versions matching our
resolver, so that is good :D.

- We also removed the app/Modes/CheckMode.hs empty file, containing a
mode that was planned but not developed (yet?).