Added parsing to a structure configuration file
Added a new step in the initialization when a file, called
"structure.ini" (an example of such a file was also included in the
project), which MUST be placed directly under the root of the proof
system, is read in order to compose the overarching ProofSystemConfig
environment configuration value for the program by the getConf function.

This structure configuration file contains information such as the
Discipline Directories of the project, their dependency allowances, the
project hidden (not recurse) directories and directory base names, and
the name of the project file (by default _CoqProject).

The reason for this change is so that it makes it easier for people to
fork the Pruvaro proof system and change some settings for personal
proof systems, without requiring an additional fork and rewrite of
Pruvarilo, as those fields were previously hardcoded to Pruvarilo.