Changed the section title for Discipline Directory sections
Chenged the case of the Discipline Directory sections in the
structure.ini file from "Discipline Directory: BaseName", where
"BaseName" is the base name of the Discipline Directory being declared,
to "DISCIPLINE DIRECTORY: BaseName". This required a modification in how
the parser looks for those sections.

Notice that the older version probably would have worked anyway, since,
according to the config-ini package's documentation, they consider the
section names in a case-insensitive way.