User warning when adding case-insenstive conflicting files
Adds a warning list, similar to the one already implemented for file
base names failing the name validity predicate, for files which have
case-insenstive base name clashes with other files in the same parent
directory. This is to help with cross-system compatibility.

As with the name validity warning, this warning does not prohibit the
user from adding new file declarations.