Replaced a rename with a copy then remove procedure
Replaced the step where the temporary file was renamed (in the more
general sense, like in the Unix mv command) from the temporary directory
to the Discipline Direcotory, replacing atomically the _CoqProject file,
with a copy (which also replaces atomically the target file, if it
exists) and remove two-step procedure.

The reason, as documented in the comments, is because the
UnliftIO.renameFile function (a wrapped-up version of the directory
package's System.Directory) may fail with an UnsupportedOperation
exception when the source (generally under /tmp) and target (usually
somewhere in the home directory) are in different file systems. I think
this is due to it being implemented with the GNU C Library function
"rename", which, according to documentation, throws an EXDEV exception
under these circumstances.