Corrects display logic when no files are added to the first DiscDir
Previously, if the first Discipline Directory in the list would not
receive any additional files (for example, if the first Discipline
Directory was the Mathematics one, and you just added a file to the
Physics Discipline Directory, which is hidden from the Mathematics one),
the program would not show any file being added at all, due to erroneous
logic.

Consequentially, no files could be added to the Physics Discipline
Directory in the current Pruvaro state.

Also corrects a comment error, where a value of a list type was being
described as a tuple type. Instead, it is a list of that tuple type.